requirements.txt
# requirements.txt for Legal RAG App
streamlit
pymupdf
langchain
google-generativeai
faiss-cpu
dotenv
sentence-transformers
